feat(docker): update Temporal worker setup and configuration
- Upgraded the Temporal Docker image to version 1.27.2 for improved features and stability.
- Refactored the Dockerfile for the Temporal worker, enhancing health check commands and environment variable management.
- Introduced a new Dockerfile for the main Temporal worker, establishing a structured build process with development and production stages.
- Added example workflows and activities, including dynamic loading in the worker entry point.
- Created a README and TypeScript configuration for the new worker structure, providing guidance for local development.

These changes enhance the Temporal worker architecture, facilitating better development practices and deployment capabilities.
