+++ workers/main/tsconfig.json
